Moved some files, removed some files, Dispatcher instead of AStinkTask
