+++ Health/src/main/java/com/sciencesquad/health/prescriptions/PrescriptionAlarm.java
@@ -5,6 +5,7 @@ import android.content.Intent;
 import android.util.Log;
 
 import com.sciencesquad.health.core.BaseApp;
+import com.sciencesquad.health.core.util.AlarmSender;
 
 /**
  * Created by andrew on 3/4/16.
