+++ Health/src/main/java/com/sciencesquad/health/sleep/SleepFragment.java
@@ -5,7 +5,6 @@ import android.graphics.Color;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v4.content.ContextCompat;
 import android.support.v7.widget.CardView;
 import android.transition.Visibility;
 import android.util.Log;
@@ -18,10 +17,10 @@ import com.github.javiersantos.materialstyleddialogs.enums.Duration;
 import com.sciencesquad.health.R;
 import com.sciencesquad.health.core.BaseFragment;
 import com.sciencesquad.health.core.Module;
-import com.sciencesquad.health.core.alarm.AlarmSender;
 import com.sciencesquad.health.core.ui.RevealTransition;
 import com.sciencesquad.health.core.util.StaticPagerAdapter;
 import com.sciencesquad.health.databinding.FragmentSleepBinding;
+import com.sciencesquad.health.core.util.AlarmSender;
 import com.wdullaer.materialdatetimepicker.time.TimePickerDialog;
 
 import java8.util.stream.Stream;
